AI Agents/langgraph-agents/memory-agent/pyproject.toml
[project]
name = "memory-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-graph-retriever>=0.8.0",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.10",
    "python-dotenv>=1.1.1",
]
